backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.100.0",
    "imbalanced-learn>=0.10.0",
    "joblib>=1.2.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.9",
    "missingno>=0.5.2",
    "numpy>=1.23.0",
    "paho-mqtt>=1.6.1",
    "pandas>=1.5.0",
    "pymodbus==3.6.8",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "requests>=2.28.0",
    "scikit-learn==1.6.1",
    "seaborn>=0.13.2",
    "uvicorn[standard]>=0.22.0",
    "xgboost>=1.7.0",
]
